+++ bin/openapi2postmanv2.js
@@ -1,16 +1,49 @@
 #!/usr/bin/env node
-var program = require('commander'),
+var _ = require('lodash'),
+  program = require('commander'),
   Converter = require('../index.js'),
   fs = require('fs'),
   path = require('path'),
+  availableOptions = require('../lib/options').getOptions('use', { usage: ['CONVERSION'] }),
   inputFile,
   outputFile,
   prettyPrintFlag,
   configFile,
+  definedOptions,
   testFlag,
   testsuiteFile,
   swaggerInput,
   swaggerData;
+
+/**
+ * Parses comma separated options mentioned in command args and generates JSON object
+ *
+ * @param {String} value - User defined options value
+ * @returns {Object} - Parsed option in format of JSON object
+ */
+function parseOptions (value) {
+  let definedOptions = value.split(','),
+    parsedOptions = {};
+
+  _.forEach(definedOptions, (definedOption) => {
+    let option = definedOption.split('=');
+
+    if (option.length === 2 && _.includes(_.keys(availableOptions), option[0])) {
+      try {
+        // parse parsable data types (e.g. boolean, integer etc)
+        parsedOptions[option[0]] = JSON.parse(option[1]);
+      }
+      catch (e) {
+        // treat value as string if can not be parsed
+        parsedOptions[option[0]] = option[1];
+      }
+    }
+    else {
+      console.warn('\x1b[33m%s\x1b[0m', 'Warning: Invalid option supplied ', option[0]);
+    }
+  });
+  return parsedOptions;
+}
 
 program
   .version(require('../package.json').version, '-v, --version')
@@ -18,9 +51,9 @@ program
   .option('-o, --output <output>', 'Write the collection to an output file')
   .option('-t, --test', 'Test the OPENAPI converter')
   .option('-p, --pretty', 'Pretty print the JSON file')
-  .option('-c, --config <config>', 'JSON file containing Converter options')
+  .option('-c, --options-config <optionsConfig>', 'JSON file containing Converter options')
+  .option('-O, --options <options>', 'comma separated list of options', parseOptions)
   .option('-g, --generate <generate>', 'Generate postman tests given the JSON file with test options');
-
 
 program.on('--help', function() {
   /* eslint-disable */
@@ -46,6 +79,7 @@ outputFile = program.output || false;
 testFlag = program.test || false;
 prettyPrintFlag = program.pretty || false;
 configFile = program.config || false;
+definedOptions = program.options || {};
 testsuiteFile = program.generate || false;
 swaggerInput;
 swaggerData;
@@ -61,14 +95,16 @@ swaggerData;
 function writetoFile(prettyPrintFlag, file, collection) {
   if (prettyPrintFlag) {
     fs.writeFile(file, JSON.stringify(collection, null, 4), (err) => {
-      if (err) { console.log('Could not write to file', err); }
-      console.log('Conversion successful', 'Collection written to file');
+      if (err) { console.log('Could not write to file', err); } // eslint-disable-line no-console
+      // eslint-disable-next-line no-console
+      console.log('\x1b[32m%s\x1b[0m', 'Conversion successful, collection written to file');
     });
   }
   else {
     fs.writeFile(file, JSON.stringify(collection), (err) => {
-      if (err) { console.log('Could not write to file', err); }
-      console.log('Conversion successful', 'Collection written to file');
+      if (err) { console.log('Could not write to file', err); } // eslint-disable-line no-console
+      // eslint-disable-next-line no-console
+      console.log('\x1b[32m%s\x1b[0m', 'Conversion successful, collection written to file');
     });
   }
 }
@@ -81,15 +117,21 @@ function writetoFile(prettyPrintFlag, file, collection) {
 function convert(swaggerData) {
   let options = {};
 
+  // apply options from config file if present
   if (configFile) {
     configFile = path.resolve(configFile);
-    console.log('Config file: ', configFile);
+    console.log('Options Config file: ', configFile); // eslint-disable-line no-console
     options = JSON.parse(fs.readFileSync(configFile, 'utf8'));
+  }
+
+  // override options provided via cli
+  if (definedOptions) {
+    options = definedOptions;
   }
 
   if (testsuiteFile) {
     testsuiteFile = path.resolve(testsuiteFile);
-    console.log('Testsuite file: ', testsuiteFile);
+    console.log('Testsuite file: ', testsuiteFile); // eslint-disable-line no-console
     options.testSuite = true;
     options.testSuiteSettings = JSON.parse(fs.readFileSync(testsuiteFile, 'utf8'));
   }